// src/models/interfaces/profile.ts
export interface ProfileProps {
    profile:{
        accountaccess: string;
        adminaccess: string;
        business: string;
        campaignaccess: string;
        carrotaccess: string;
        carrotrole: string;
        commsaccess: string;
        contentaccess: string;
        customercare: string;
        email: string;
        firstname: string;
        id: 0,
        isactive: string;
        lastname: string;
        loginid: string;
        mobile: string;
        offersaccess: string;
        reporting: string;
        scratchaccess: string;
        status: string;
        superadmin: string;
        sysadmin: string;
    }
   
}